CFS-83:Rebase And Apply Left changes
